API call for matches now in MainWrapper, as to feed sidebar too
